+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [UNRELEASED]
+## [Unreleased]
 
 ### Fixed
 
+- Fix crash when importing an ISO8601 datetime value with microseconds/timezone
 - Fix : confirm dialog js escaping
 
 ## [2.15.8] - 2026-06-24